Resolved review comments
